requirements.txt
fastapi
uvicorn
pydantic
requests
PyYAML
openai
